requirements.txt
# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# HTTP client
httpx>=0.26.0

# Retry logic
tenacity>=8.2.0

# Monitoring
sentry-sdk[fastapi]>=1.40.0

# Supabase
supabase>=2.3.0
python-jose[cryptography]>=3.3.0

# LLM
anthropic>=0.40.0

# Data
pandas>=2.1.0

# SSE streaming
sse-starlette>=1.8.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-mock>=3.12.0
